+++ zulipterminal/platform_code.py
@@ -1,8 +1,11 @@
 import platform
+import subprocess
 
+import lxml.html
 from typing_extensions import Literal
 
 
+# PLATFORM DETECTION
 SupportedPlatforms = Literal["Linux", "MacOS", "WSL"]
 AllPlatforms = Literal[SupportedPlatforms, "unsupported"]
 
@@ -16,3 +19,37 @@ elif raw_platform == "Darwin":
     PLATFORM = "MacOS"
 else:
     PLATFORM = "unsupported"
+
+
+# PLATFORM DEPENDENT HELPERS
+def notify(title: str, html_text: str) -> str:
+    document = lxml.html.document_fromstring(html_text)
+    text = document.text_content()
+
+    command_list = None
+    if PLATFORM == "MacOS":
+        command_list = [
+            "osascript",
+            "-e",
+            "on run(argv)",
+            "-e",
+            "return display notification item 1 of argv with title "
+            'item 2 of argv sound name "ZT_NOTIFICATION_SOUND"',
+            "-e",
+            "end",
+            "--",
+            text,
+            title,
+        ]
+    elif PLATFORM == "Linux":
+        command_list = ["notify-send", "--", title, text]
+
+    if command_list is not None:
+        try:
+            subprocess.run(
+                command_list, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL
+            )
+        except FileNotFoundError:
+            # This likely means the notification command could not be found
+            return command_list[0]
+    return ""
